Extract doChores()
